+++ http-client/test-nonet/Network/HTTP/Client/RequestSpec.hs
@@ -54,6 +54,25 @@ spec = do
         field `shouldSatisfy` isJust
         field `shouldBe` Just "Basic dXNlcjpwYXNz"
 
+    describe "getUri" $ do
+      context "when protocol is http and port is 80" $ do
+        it "omits port" $ do
+          let url = "http://example.com/"
+          request <- parseRequest url
+          show (getUri request) `shouldBe` url
+
+      context "when protocol is https and port is 443" $ do
+        it "omits port" $ do
+          let url = "https://example.com/"
+          request <- parseRequest url
+          show (getUri request) `shouldBe` url
+
+      context "when protocol is https and port is 80" $ do
+        it "does not omit port" $ do
+          let url = "https://example.com:80/"
+          request <- parseRequest url
+          show (getUri request) `shouldBe` url
+
     describe "Show Request" $
       it "redacts authorization header content" $ do
         let request = defaultRequest { requestHeaders = [("Authorization", "secret")] }